Script fails to find protobuf version

+++ protobuf.sh
@@ -2,6 +2,7 @@
 # Author: Connor Weeks
 
 SCRIPT_DIR="$(cd "$(dirname "${0}")"; pwd)"
+NUMJOBS=${NUMJOBS:-1}
 RED="\033[1;31m"
 YELLOW="\033[1;33m"
 GREEN="\033[0;32m"
@@ -22,7 +23,7 @@ print_usage () {
 |
 |     Generates the cmake files for the given installation of tensorflow
 |     and writes them to <cmake-dir>. If 'generate installed' is executed,
-|     <install-dir> corresponds to the directory Protobuf was installed to; 
+|     <install-dir> corresponds to the directory Protobuf was installed to;
 |     defaults to /usr/local.
 |
 | --> ${0} install <tensorflow-source-dir> [<install-dir> <download-dir>]
@@ -56,17 +57,17 @@ find_protobuf () {
     QUOTE_START="\s*=\s*\\\""
     QUOTE_END="\\\"\s*,\s*"
     FOOTER="\)"
-    PROTOBUF_NAME="${NAME_START}protobuf${QUOTE_END}"
-    
-    
-    
+    PROTOBUF_NAME="${NAME_START}com_google_protobuf${QUOTE_END}"
+
+
+
     if [ ${1} -eq 0 ]; then
 	PROTOBUF_REGEX="${HTTP_HEADER}${ANY}${PROTOBUF_NAME}${ANY}${FOOTER}"
 	FOLDER="s/\s*strip_prefix${QUOTE_START}\(${ANY_NO_QUOTES}\)${QUOTE_END}/\1/p"
-	
+
 	URL="s/\s*url${QUOTE_START}\(${ANY_NO_QUOTES}\)${QUOTE_END}/\1/p"
-	
-	
+
+
 	PROTOBUF_TEXT=$(grep -Pzo ${PROTOBUF_REGEX} ${TF_DIR}/tensorflow/workspace.bzl) || fail
 	PROTOBUF_URL=$(echo "${PROTOBUF_TEXT}" | sed -n ${URL})
 	PROTOBUF_URLS[0]=${PROTOBUF_URL}
@@ -191,44 +192,46 @@ if [ "${MODE}" == "install" ]; then
 	    # download protobuf from http archive
 	    cd ${DOWNLOAD_DIR} || fail
 	    wget ${URL} && FOUND_URL=1 && break
-	    
+
 	fi
     done
     if [ ${FOUND_URL} -eq 0 ]; then
 	echo "${RED}Could not download Protobuf${NO_COLOR}"
 	exit 1
     fi
-    
+
     tar -xf ${PROTOBUF_ARCHIVE} || fail
     cd ${PROTOBUF_FOLDER} || fail
-    
-    
-    
-    
+
+
+
+
     # configure
     ./autogen.sh || fail
     ./configure --prefix=${INSTALL_DIR} || fail
     echo "Starting protobuf install."
     # build and install
-    make || fail
+    make -j$NUMJOBS || fail
     make check || fail
     make install || fail
-    ldconfig || fail
+    if [ `id -u` == 0 ]; then
+        ldconfig || fail
+    fi
     cd ${DOWNLOAD_DIR} || fail
     rm ${PROTOBUF_ARCHIVE} || fail
     echo "Protobuf has been installed to ${INSTALL_DIR}"
 elif [ "${MODE}" == "generate" ]; then
-    
+
     if [ "${GENERATE_MODE}" == "installed" ]; then
 	# try to locate protobuf in INSTALL_DIR
 	if [ -d "${INSTALL_DIR}/include/google/protobuf" ]; then
             echo -e "${GREEN}Found Protobuf in ${INSTALL_DIR}${NO_COLOR}"
 	else
- 	    echo -e "${YELLOW}Warning: Could not find Protobuf in ${INSTALL_DIR}${NO_COLOR}"	
+ 	    echo -e "${YELLOW}Warning: Could not find Protobuf in ${INSTALL_DIR}${NO_COLOR}"
 	fi
     fi
-    
-    PROTOBUF_OUT="${CMAKE_DIR}/Protobuf_VERSION.cmake"	
+
+    PROTOBUF_OUT="${CMAKE_DIR}/Protobuf_VERSION.cmake"
     echo "set(Protobuf_URL ${PROTOBUF_URL})" > ${PROTOBUF_OUT} || fail
     if [ "${GENERATE_MODE}" == "external" ]; then
 	echo "Wrote Protobuf_VERSION.cmake to ${CMAKE_DIR}"
